Refactor: unify ls/ftp-ls code
Both the 'ls' command and the 'ls' subcommand of the 'ftp' command use
some code that is very similar. Modifications must be done in two places
instead of one and this can lead to inconsistencies.

Refactor code so that both paths use the same function with the same opt
spec.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>