Update bupsplit.[ch] to have a less restrictive BSD-style license.
Nobody else owns any copyright on those files, so I can do this without
asking anyone else for permission :)

Some people have asked to use these files in their own non-GPL/LGPL
projects, and although the LGPL permits this, the files are so small that
there's no reason to be obnoxious about it.  So let's just let them do it
the easy way.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>