test-compression.sh: use "tar ... | wc -c" instead of du.
This test was failing on a ZFS system with compression enabled because
du wasn't summing the actual file sizes.  To fix that, use tar/wc so
that we'll calculate the actual tree size without having to worry
about the portability of du's --apparent-size argument.

Thanks to Björn Seifert <seifert@oern.de> for reporting the problem
and Greg Troxel <gdt@lexort.com> for suggesting this fix.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>