index: Correct inconsitency in documentation of modes
In the optspec, --clear and --check are documented as being modes, but
in the man page they are simple options.

Also, even though --clear and --check are listed as modes in the
optspec, they don't appear in the synopsis line.

According to the code, --clear and --check do behave as modes so we
should be consistent in how we document them.

Signed-off-by: Gabriel Filion <gabster@lelutin.ca>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>